refactor(radix): one file per component, matching mantine
chrome.tsx and tables.tsx split into Toolbar, ActiveFilterChips,
BulkActionBar, PaginationFooter, ErrorState, TableSkeleton, FilterDrawer,
ExpandToggle, RowActionButtons, DesktopTable and MobileCards — the layout
mantine already uses, so a component is found under its own name in any
adapter. No behaviour, exports or public API change.
